Modifications integrating Tamer with Compiled Tetris
Reworked StateEvaluation
Pulled simulation function from Tetris.py and put it in TetrisSym.py
Trimmed Tetris.py
