+++ Tetris/Tetris.py
@@ -6,20 +6,14 @@
 # Original code from https://levelup.gitconnected.com/writing-tetris-in-python-2a16bddb5318
 
 import os
-import csv
 from typing import Counter
 import pygame
 import random
 import platform
 import sys
 import re
-import numpy
-import copy
 import time
-import tensorflow as tf
 import StateEvaluation
-from tensorflow.keras.models import Model
-from tensorflow.keras.layers import Dense, Input
 
 # This is an optional import that allows you to switch panels with the number keys (Windows Only)
 try:
@@ -27,18 +21,18 @@ try:
 except ImportError:
     print("Could not import pywinauto")
 
-checkPointPath = "tamer.hdf5"
-game_speed_modifier = 0.25
-upper_speed_bound = 0.01
-queue_size = 4
-Is_Master = False
-Should_Load_Model = False
-Activate_Hidden_Rule = False
+game_speed_modifier   = 0.25
+upper_speed_bound     = 0.01
+queue_size            = 4
+game_id               = 0
+Is_Master             = False
+Should_Load_Model     = False
+Activate_Hidden_Rule  = False
 Activate_Hidden_Piece = False
 Activate_Hidden_Delay = 60
-Speed_Increase = False
+Speed_Increase        = False
 hidden_piece_timer_elapsed = False
-game_id = 0
+
 pygame.init()
 
 # Use the number keys 0-9 to toggle between windows
@@ -168,163 +162,36 @@ else:
 
 Read_Config()
 
-## CURRENT BASIC TAMER IMPLEMENTATION
-class Tamer:
-    def __init__(self):
-        self.compiled = False
-        self.model = None
-        self.runRandom = False
-        self.compileModel()
-        self.record = []
-        self.arecord = []
-        self.frecord = []
-        self.load_record = []
-        self.load_frecord = []
-
-    def compileModel(self, optimizer=None, learning_rate=0.001, metrics=[]):
-        # construct generic model if one is not specified
-        # initializer = tf.keras.initializers.Zeros()
-        if self.model == None:
-            input_shape = 46
-            input1 = Input(shape=(input_shape,))
-            x1 = Dense(1, kernel_initializer="zeros", bias_initializer="zeros")(input1)
-            inputs = [input1]
-
-            self.model = Model(inputs, outputs=[x1])  # ,y2,y3])
-
-        # compile the model
-        if optimizer == None:
-            self.model.compile(
-                optimizer=tf.keras.optimizers.SGD(lr=0.000005 / 47.0), loss="mse"
-            )
-        else:
-            self.model.compile(optimizer=optimizer, loss="mse")
-        self.compiled = True
-
-    def load_weights(self, filepath):
-        if self.compiled == True:
-            self.model.load_weights(filepath)
-
-    def save_weights(self, filepath):
-        if self.compiled == True:
-            self.model.save_weights(filepath)
-
-    def forward(self, state_state_feats):
-        action = numpy.random.randint(0, len(state_state_feats))
-        ssfeats = numpy.zeros([len(state_state_feats), 46])
-        for s in range(len(state_state_feats)):
-            for f in range(46):
-                ssfeats[s, f] = state_state_feats[s][f]
-        pred_rewards = numpy.zeros(len(state_state_feats))
-
-        rid = numpy.random.randint(0, 10)
-
-        if rid >= 8 and self.runRandom:
-            action = numpy.random.randint(0, len(state_state_feats))
-        else:
-            if self.compiled == True:
-                pred_rewards = self.model.predict(ssfeats, verbose=0)
-                action = numpy.argmax(pred_rewards)
-            else:
-                action = numpy.random.randint(0, len(state_state_feats))
-
-        return action, pred_rewards
-
-    def backward(self, prev_state_state_feats, prev_action, human_reward):
-        # do a gradient update step
-        if (
-            human_reward != 0
-            and prev_state_state_feats is not None
-            and prev_action is not None
-            and self.compiled == True
-        ):
-            x = numpy.reshape(
-                prev_state_state_feats[prev_action],
-                (1, len(prev_state_state_feats[prev_action])),
-            ).astype("float32")
-            y = numpy.reshape(numpy.array(human_reward, dtype="float32"), (1))
-            self.model.fit(x, y, verbose=0)
-
-    def batch_backward(self):
-        if len(self.record) > 0:
-            ssfeats = numpy.zeros([len(self.record), 46])
-            yvalue = numpy.zeros([len(self.frecord)])
-            for s in range(len(self.record)):
-                tmp = self.record[s]
-                feat = tmp[self.arecord[s]]
-                yvalue[s] = self.frecord[s]
-                for f in range(46):
-                    ssfeats[s, f] = feat[f]
-            self.model.fit(ssfeats, yvalue, verbose=0)
-
-    def all_backward(self):
-        ssfeats = numpy.zeros([len(self.load_record), 46])
-        yvalue = numpy.zeros([len(self.load_frecord)])
-        for s in range(len(self.load_record)):
-            tmp = self.load_record[s]
-            # feat      = tmp[self.load_arecord[s]]
-            yvalue[s] = self.load_frecord[s]
-            for f in range(46):
-                ssfeats[s, f] = tmp[f]
-        self.model.fit(ssfeats, yvalue, verbose=0)
-
-    def load_data(self, filepath):
-        # LOOP ALL TEXTFILES (.CSV) IN DIRECTORY
-        testFiles = os.listdir(filepath)
-        self.load_record = []
-        self.load_frecord = []
-        for t in range(len(testFiles)):
-            with open(filepath + testFiles[t], mode="r") as csv_file:
-                csv_reader = csv.reader(csv_file)
-                ndata = list(csv_reader)
-                for s in range(len(ndata)):
-                    row = ndata[s]
-                    self.load_record.append(copy.deepcopy(row[0 : (len(row) - 1)]))
-                    self.load_frecord.append(copy.deepcopy(row[-1]))
-        self.all_backward()
-
-    def save_data(self, filename):
-        ssfeats = numpy.zeros([len(self.record), 46])
-        yvalue = numpy.zeros([len(self.frecord)])
-        for s in range(len(self.record)):
-            tmp = self.record[s]
-            feat = tmp[self.arecord[s]]
-            yvalue[s] = self.frecord[s]
-            for f in range(46):
-                ssfeats[s, f] = feat[f]
-        textfile = open(filename, "w")
-        for s in range(len(self.record)):
-            for f in range(46):
-                textfile.write(str(ssfeats[s, f]) + ",")
-            textfile.write(str(yvalue[s]) + "\n")
-        textfile.close()
-
 class Tetris:
     level = 2
     score = 0
     state = "start"
     field = []
-    finalStates = []
-    actionTree = []
+    finalStates   = []
+    actionTree    = []
     visitedStates = []
     height = 0
-    width = 0
-    x = 100
-    y = 60
-    zoom = 20
+    width  = 0
+    x      = 100
+    y      = 60
+    zoom   = 20
     figure = None
+    
     figure_queue = []
-    reward = 0
-    sim = False
-    should_flash_reward_text = False
-    reward_text_flash_time = 90
+    reward       = 0
+    sim          = False
+    
+    should_flash_reward_text  = False
+    reward_text_flash_time    = 90
     reward_text_flash_counter = 0
-    feedback = 0
-    newFig = 0
-    lastMove = -1
-    gameid = 0
-    numGames = 0
+    feedback  = 0
+    newFig    = 0
+    lastMove  = -1
+    gameid    = 0
+    numGames  = 0
     numPieces = 0
+    playAI = False
+    trainAI = False
 
     def __init__(self, height, width):
         self.height = height
@@ -334,18 +201,6 @@ class Tetris:
         self.reward = 0
         self.state = "start"
         self.numPieces = 0
-
-        # Hand-crafted Tetris Features (for TAMER)
-        self.NUM_FEATS = 46
-        self.COL_HT_START_I = 0
-        self.MAX_COL_HT_I = 10
-        self.COL_DIFF_START_I = 11
-        self.NUM_HOLES_I = 20
-        self.MAX_WELL_I = 21
-        self.SUM_WELL_I = 22
-        self.SQUARED_FEATS_START_I = 23
-        self.SCALE_ALL_SQUARED_FEATS = False
-        self.HT_SQ_SCALE = 100.0
 
         for i in range(height):
             new_line = []
@@ -373,36 +228,6 @@ class Tetris:
                     ):
                         intersection = True
         return intersection
-
-    ## SIMULATION FUNCTION
-    def intersectsSym(self, board, figure):
-        intersection = False
-        for i in range(4):
-            for j in range(4):
-                if i * 4 + j in figure.image():
-                    if (
-                        i + figure.y > self.height - 1
-                        or j + figure.x > self.width - 1
-                        or j + figure.x < 0
-                        or board[i + figure.y][j + figure.x] > 0
-                    ):
-                        intersection = True
-        return intersection
-
-    ## SIMULATION FUNCTION
-    def break_linesSym(self, board):
-        lines = 0
-        for i in range(1, self.height):
-            zeros = 0
-            for j in range(self.width):
-                if board[i][j] == 0:
-                    zeros += 1
-            if zeros == 0:
-                lines += 1
-                for i1 in range(i, 1, -1):
-                    for j in range(self.width):
-                        board[i1][j] = board[i1 - 1][j]
-        return board
 
     def break_lines(self):
         lines = 0
@@ -433,21 +258,6 @@ class Tetris:
         self.figure.y -= 1
         self.freeze()
 
-    ## SIMULATION FUNCTION
-    def go_downSym(self, board, figure):
-        fig = copy.deepcopy(figure)
-        if fig.x != figure.x:
-            print("COPY ERROR")
-            time.sleep(10)
-        brd = copy.deepcopy(board)
-        fig.y += 1
-        symStopped = False
-        if self.intersectsSym(brd, fig):
-            fig.y -= 1
-            brd = self.freezeSym(brd, fig)
-            symStopped = True
-        return brd, fig, symStopped
-
     def go_down(self):
         game.newFig = 0
 
@@ -467,24 +277,6 @@ class Tetris:
         if self.intersects():
             self.state = "gameover"
 
-    ## SIMULATION FUNCTION
-    def freezeSym(self, board, figure):
-        for i in range(4):
-            for j in range(4):
-                if i * 4 + j in figure.image():
-                    board[i + figure.y][j + figure.x] = figure.color
-        board = self.break_linesSym(board)
-        return board
-
-    def go_sideSym(self, dx, board, figure):
-        fig = copy.deepcopy(figure)
-
-        old_x = figure.x
-        fig.x += dx
-        if self.intersectsSym(board, fig):
-            fig.x = old_x
-        return fig
-
     def go_side(self, dx):
         game.newFig = 0
 
@@ -497,15 +289,6 @@ class Tetris:
         if self.intersects():
             self.figure.x = old_x
 
-    def rotateSym(self, board, figure):
-        fig = copy.deepcopy(figure)
-        old_rotation = figure.rotation
-        fig.rotate()
-        if self.intersectsSym(board, fig):
-            fig.rotation = old_rotation
-
-        return fig
-
     def rotate(self):
         game.newFig = 0
 
@@ -515,141 +298,6 @@ class Tetris:
         self.figure.rotate()
         if self.intersects():
             self.figure.rotation = old_rotation
-
-    def getFeatures(self, board):
-        featsArray = (
-            numpy.zeros(
-                46,
-            )
-            - 1
-        )
-        featsArray[self.NUM_HOLES_I] = 0
-        featsArray[self.SUM_WELL_I] = 0
-        board = numpy.array(board)
-
-        for row in range(self.height + 1):
-            for col in range(self.width):
-                if board[row][col] > 0:  # FILLED CELL
-                    if featsArray[self.COL_HT_START_I + col] == -1:
-                        featsArray[self.COL_HT_START_I + col] = row
-                    if featsArray[self.MAX_COL_HT_I] == -1:
-                        featsArray[self.MAX_COL_HT_I] = row
-                else:  # empty cell
-                    if featsArray[self.COL_HT_START_I + col] != -1:
-                        featsArray[self.NUM_HOLES_I] += 1
-        if featsArray[self.MAX_COL_HT_I] == -1:
-            featsArray[self.MAX_COL_HT_I] = self.rows
-
-        # get column difference features
-        for col in range(self.width - 1):
-            featsArray[self.COL_DIFF_START_I + col] = abs(
-                featsArray[self.COL_HT_START_I + col]
-                - featsArray[self.COL_HT_START_I + col + 1]
-            )
-
-        # get well depth features
-        for col in range(self.width):
-            wellDepth = self.getWellDepth(col, board)
-            featsArray[self.SUM_WELL_I] += wellDepth
-            if wellDepth > featsArray[self.MAX_WELL_I]:
-                featsArray[self.MAX_WELL_I] = wellDepth
-
-        # get squared features and scale them so they're not too big
-        for i in range(self.SQUARED_FEATS_START_I):
-            featsArray[self.SQUARED_FEATS_START_I + i] = numpy.square(featsArray[i])
-            if i <= self.MAX_COL_HT_I or self.SCALE_ALL_SQUARED_FEATS:
-                featsArray[self.SQUARED_FEATS_START_I + i] /= self.HT_SQ_SCALE
-
-        return featsArray
-
-    def getWellDepth(self, col, board):
-        depth = 0
-        for row in range(self.height):
-            if board[row][col] > 0:  # encountered a filled space, stop counting
-                return depth
-            else:
-                if (
-                    depth > 0
-                ):  # if well-depth count has begun, dont require left or right side to be filled
-                    depth += 1
-                # check if both the cell to the left if full and if the cell to the right is full
-                elif (col == 0 or board[row][col - 1] > 0) and (
-                    col == self.width - 1 or board[row][col + 1] > 0
-                ):
-                    depth += 1
-        return depth
-
-    def forwardProject(self, board, figure, alist):
-        nlist = copy.deepcopy(alist)
-        brd = copy.deepcopy(board)
-        fig = copy.deepcopy(figure)
-        x = copy.deepcopy(fig.x)
-        y = copy.deepcopy(fig.y)
-        r = copy.deepcopy(fig.rotation)
-
-        for m in range(0, len(self.visitedStates)):
-            vs = self.visitedStates[m]
-            if x == vs[0] and y == vs[1] and r == vs[2]:
-                return
-
-        if len(self.finalStates) > 50:
-            return
-        for a in range(0, 4):
-            symStopped = False
-            # TAKE ACTION A
-            if a == 0:
-                nboard, nfigure, symStopped = self.go_downSym(brd, fig)
-                nlist.append(a)
-            elif a == 1:
-                nfigure = self.go_sideSym(1, brd, fig)
-                nboard, nfigure, symStopped = self.go_downSym(brd, nfigure)
-                nlist.append(a)
-                nlist.append(0)
-            elif a == 2:
-                nfigure = self.go_sideSym(-1, brd, fig)
-                nboard, nfigure, symStopped = self.go_downSym(brd, nfigure)
-                nlist.append(a)
-                nlist.append(0)
-            elif a == 3:
-                nfigure = self.rotateSym(brd, fig)
-                nboard, nfigure, symStopped = self.go_downSym(brd, nfigure)
-                nlist.append(a)
-                nlist.append(0)
-
-            if symStopped == False:
-                self.forwardProject(nboard, nfigure, nlist)
-                nlist.pop()
-                if a > 0:
-                    nlist.pop()
-            else:
-                match = False
-                for s in range(0, len(self.finalStates)):
-                    s0 = self.finalStates[s]
-                    matched = True
-                    for i in range(0, self.height):
-                        for j in range(0, self.width):
-                            if s0[i][j] != nboard[i][j]:
-                                matched = False
-                                break
-                        if matched == False:
-                            break
-                    if matched == True:
-                        match = True
-                        break
-                if match == False:
-                    self.finalStates.append(copy.deepcopy(nboard))
-                    self.actionTree.append(copy.deepcopy(nlist))
-                # time.sleep(1.0)
-                nlist.pop()
-                if a > 0:
-                    nlist.pop()
-
-            if len(nlist) > 25:
-                # print("POPPING OUT")
-                return
-
-        self.visitedStates.append([x, y, r])
-        return
 
     def update_reward(self, score_to_add):
         self.reward += score_to_add
@@ -739,34 +387,18 @@ counter = 0
 pressing_down = False
 last_move = ""
 auto_restart = False
-tamer = Tamer()
-#tamer.load_weights(checkPointPath)
-game.gameid = game_id
-gameSym = Tetris(20, 10)
-counter = 0
+
+game.gameid   = game_id
+counter       = 0
 pressing_down = False
-last_move = ""
-auto_restart = True
-playAI = False
-trainAI = True
-rewardLearn = False
-gameCounter = 0
-# lastWrite = 0
-# version = 6
-prevx = 0
-prevy = 0
-actLoc = 0
-prev_feedback = 0
-prev_prev_feedback = 0
-prev_reward = 0
-prev_prev_reward = 0
-prev_action = None
-prev_state_state_feats = None
-prev_prev_state_state_feats = None
-state_state_feats = None
-action = 0
-runQuick = False
-Q_PLAN = False
+last_move     = ""
+auto_restart  = True
+rewardLearn   = False
+gameCounter   = 0
+
+prevx      = 0
+prevy      = 0
+runQuick   = False
 game_stats = []
 
 def Find_Area(game):
@@ -790,211 +422,84 @@ while not done:
     if counter > 100000:
         counter = 0
 
-    if playAI:
-        if game.newFig == 1:
-            # LEARN FROM LAST EPISODE FIRST!!!
-            if trainAI and game.feedback != 0:
-                prev_state_state_feats = copy.deepcopy(state_state_feats)
-                prev_action = copy.deepcopy(action)
-                tamer.backward(prev_state_state_feats, prev_action, game.feedback)
-                tamer.record.append(prev_state_state_feats)
-                tamer.arecord.append(prev_action)
-                tamer.frecord.append(game.feedback)
-
-            prev_state_state_feats = copy.deepcopy(state_state_feats)
-            prev_action = copy.deepcopy(action)
-            # TAMER LOOP
-            # STEP 1: DEEP COPY OF GAME
-            gameSym = copy.deepcopy(game)
-
-            # STEP 2: FORWARD PROJECT OF PIECES
-            gameSym.finalStates = []
-            gameSym.actionTree = []
-            gameSym.visitedStates = []
-            field = copy.deepcopy(gameSym.field)
-            figure = copy.deepcopy(gameSym.figure)
-            gameSym.forwardProject(field, figure, [])
-
-            # STEP 2: COMPUTE STATE
-            board = [[0 for x in range(10)] for y in range(20)]
-            board += [[1 for x in range(10)]]
-            for i in range(10):
-                for j in range(20):
-                    board[j][i] = game.field[j][i]
-
-            state0_feat = gameSym.getFeatures(board)
-
-            # STEP 3: DIFF STATES
-            state1_feats = []
-            for piecePlacement in gameSym.finalStates:
-                board = [[0 for x in range(10)] for y in range(20)]
-                board += [[1 for x in range(10)]]
-                for i in range(10):
-                    for j in range(20):
-                        board[j][i] = piecePlacement[j][i]
-                state1_feats.append(gameSym.getFeatures(board))
-
-            # get state_state features
-            state_state_feats = []
-            for state1_feat in state1_feats:
-                state_state_feats.append(state1_feat - state0_feat)
-
-            # use tamer model to get best action
-            action, actionSet = tamer.forward(state_state_feats)
-
-            finalStates = copy.deepcopy(gameSym.finalStates)
-            actionTree = copy.deepcopy(gameSym.actionTree)
-
-            if Q_PLAN:
-                # QUEUE PROJECT
-                # FOR EACH ACTION --> RUN ANOTHER PASS OF FORWARD PROJECT
-                ids = numpy.argsort(actionSet, axis=0)
-                ids = numpy.flip(ids)
-                total = 8
-                if total > len(finalStates):
-                    total = len(finalStates)
-                for f in range(total):
-                    gameSym.finalStates = []
-                    gameSym.actionTree = []
-                    gameSym.visitedStates = []
-                    idF = ids[f, 0]
-                    field = copy.deepcopy(finalStates[idF])
-                    figure = copy.deepcopy(gameSym.figure_queue[0])
-                    figure.x = gameSym.figure.x
-                    figure.y = gameSym.figure.y
-                    gameSym.forwardProject(field, figure, [])
-
-                    field = copy.deepcopy(finalStates[idF])
-                    board = [[0 for x in range(10)] for y in range(20)]
-                    board += [[1 for x in range(10)]]
-                    for i in range(10):
-                        for j in range(20):
-                            board[j][i] = field[j][i]
-
-                    stateX_feat = gameSym.getFeatures(board)
-
-                    # STEP 3: DIFF STATES
-                    stateY_feats = []
-                    for piecePlacement in gameSym.finalStates:
-                        board = [[0 for x in range(10)] for y in range(20)]
-                        board += [[1 for x in range(10)]]
-                        for i in range(10):
-                            for j in range(20):
-                                board[j][i] = piecePlacement[j][i]
-                        stateY_feats.append(gameSym.getFeatures(board))
-
-                    # get state_state features
-                    stateZ_feats = []
-                    for stateY_feat in stateY_feats:
-                        stateZ_feats.append(stateY_feat - stateX_feat)
-                    qaction, qactionSet = tamer.forward(stateZ_feats)
-                    actionSet[idF] += numpy.max(qactionSet)
-
-                action = numpy.argmax(actionSet)
-
-            # STEP 4: DETERMINE COA
-            coa = []
-            if action < len(actionTree):
-                coa = actionTree[action]
-            actLoc = 0
-
-            # UPDATE PREV STATE VALUES (TO USE ON NEXT ITERATION)
-            prev_reward = copy.deepcopy(game.score)
-            prev_prev_feedback = copy.deepcopy(prev_feedback)
-            prev_feedback = copy.deepcopy(game.feedback)
-            game.feedback = 0
-
-        if actLoc < len(coa):
-            action1 = coa[actLoc]
-            actLoc += 1
-            if action1 == 1:
-                game.go_side(1)
-            elif action1 == 2:
-                game.go_side(-1)
-            elif action1 == 3:
-                game.rotate()
-            else:
-                game.go_down()
-        else:
-            game.go_down()
-        if runQuick == False:
+    # GET COPY OF EVENTS
+    events = pygame.event.get()
+    StateEvaluation.GameStateEvaluation(game,events)
+    
+    ## FIX THE CLOCK
+    if runQuick == False:
+        if game.playAI:
             time.sleep(game_speed_modifier)
-            if Speed_Increase:
-                # Slowly increase the speed
-                game_speed_modifier *= 0.99
-                if game_speed_modifier < upper_speed_bound:
-                    game_speed_modifier = upper_speed_bound
+        if Speed_Increase:
+            # Slowly increase the speed
+            game_speed_modifier *= 0.99
+            if game_speed_modifier < upper_speed_bound:
+                game_speed_modifier = upper_speed_bound
+        
     prevx = game.figure.x
     prevy = game.figure.y
-
-    if playAI == False:
+    
+    if game.playAI == False:
         if counter % (fps // game.level // 2) == 0 or pressing_down:
             if game.state == "start":
                 game.go_down()
 
     # Listens for keypresses and calls their respective functions
-    for event in pygame.event.get():
+    for event in events:
         if event.type == pygame.QUIT:
             done = True
+        
         if event.type == pygame.KEYDOWN:
             if event.key == pygame.K_UP:
                 game.rotate()
+                game.playAI = False
+    
             if event.key == pygame.K_DOWN:
                 pressing_down = True
-                last_move = "down"
+                last_move  = "down"
+                game.playAI = False
+                
             if event.key == pygame.K_LEFT:
                 game.go_side(-1)
                 last_move = "left"
+                game.playAI = False
+                
             if event.key == pygame.K_RIGHT:
                 game.go_side(1)
                 last_move = "right"
+                game.playAI = False
+                
             if event.key == pygame.K_SPACE:
                 game.go_space()
-                last_move = "down"
+                last_move   = "down"
+                game.playAI = False
+                
             if event.key == pygame.K_ESCAPE:
-                game.newFig = 0
+                game.newFig   = 0
                 game.lastMove = 0
                 game.__init__(20, 10)
                 number_of_games_played += 1
                 game.numGames = game.numGames + 1
                 last_move = "restart"
+            
             if event.key == pygame.K_j:
                 game.encourage(1)
-                game.reward += 1
                 last_move = "encourage"
+            
             if event.key == pygame.K_k:
                 game.encourage(-1)
-                game.reward -= 1
                 last_move = "discourage"
+            
             if event.key == pygame.K_a:
-                tamer.load_data("data\\")
-                last_move = "load data"
-            if event.key == pygame.K_s:
-                tamer.save_data("data\\dataRun" + str(game.gameid) + ".csv")
-                last_move = "save data"
-            if event.key == pygame.K_w:
-                tamer.save_weights("model\\tamer" + str(game.gameid) + ".hdf5")
-            if event.key == pygame.K_b:
-                tamer.batch_backward()
-                last_move = "Tamer batch backwards"
+                game.playAI = True
+                
             if event.key == pygame.K_q:
                 if runQuick == True:
                     runQuick = False
                 else:
                     runQuick = True
                 last_move = "Toggle Speed"
-            if event.key == pygame.K_r:
-                if tamer.runRandom == True:
-                    tamer.runRandom = False
-                    print("Random OFF")
-                else:
-                    tamer.runRandom = True
-                    print("Random ON")
-            if event.key == pygame.K_t:
-                if Q_PLAN == True:
-                    Q_PLAN = False
-                else:
-                    Q_PLAN = True
+            
             if event.key == pygame.K_g:
                 textfile = open("gameStats" + str(game.gameid) + ".csv", "w")
                 for s in range(len(game_stats)):
@@ -1012,6 +517,7 @@ while not done:
                 )
                 textfile.close()
             # Used number keys to switch panels if they exist
+            
             if event.key == pygame.K_0 and platform.system() == "Windows":
                 Set_Focus(0)
             if event.key == pygame.K_1 and platform.system() == "Windows":
@@ -1075,11 +581,11 @@ while not done:
                         ],
                     )
 
-    font = pygame.font.SysFont("Calibri", 25, True, False)
+    font  = pygame.font.SysFont("Calibri", 25, True, False)
     font1 = pygame.font.SysFont("Calibri", 65, True, False)
-    text = font.render("Score: " + str(game.score), True, BLACK)
-    text_game_over = font1.render("Game Over", True, (255, 125, 0))
-    text_game_over1 = font1.render("Press ESC", True, (255, 215, 0))
+    text  = font.render("Score: " + str(game.score), True, BLACK)
+    text_game_over        = font1.render("Game Over", True, (255, 125, 0))
+    text_game_over1       = font1.render("Press ESC", True, (255, 215, 0))
     text_last_button_used = font.render(last_move, True, (0, 0, 0))
 
     if (
@@ -1119,7 +625,7 @@ while not done:
     screen.blit(text, [0, 0])
     screen.blit(reward_text, [0, 25])
     screen.blit(text_last_button_used, [0, 50])
-    StateEvaluation.GameStateEvaluation(game)
+    
     pygame.display.flip()
     clock.tick(fps)
 pygame.quit()
